backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pusher
python-dotenv
gunicorn
whitenoise
transformers
torch